% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{as_os_style_path}
\alias{as_os_style_path}
\title{Return the raw file path based on the os path system.}
\usage{
as_os_style_path(path)
}
\arguments{
\item{path}{Path for path translation}
}
\value{
path with replace the path separater if it is required.
}
\description{
Return the raw file path based on the os path system.
}
