% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{assert}
\alias{assert}
\title{Assert}
\usage{
assert(condition, message = "assert error")
}
\arguments{
\item{condition}{Condition for assert.}

\item{message}{Exception message if the condition is not satisfied.}
}
\description{
Assert: throw exception if the condition is not satisfied.
}
