% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{cache}
\alias{cache}
\title{Cache expression result}
\usage{
cache(expr, path, expire = NULL, required_ver = NULL)
}
\arguments{
\item{expr}{expression for caching}

\item{path}{path for cache file}

\item{expire}{diftim until the cache is expired, e.g. as.difftime function with unit argument.}

\item{required_ver}{integer for controlling file format version; smaller version file is always ignored.}
}
\value{
loaded file
}
\description{
expression result is saved as rds with time stamp. If cached file already exist, it try to use it.
}
