% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{callstack_info}
\alias{callstack_info}
\title{Create callstack data at the point this function is called}
\usage{
callstack_info(IgnoreLevel = 0)
}
\arguments{
\item{IgnoreLevel}{Number of callstack data level which should be ignored from return value}
}
\value{
callstack dataframe with name, line and filename.
}
\description{
Create callstack data from sys.calls function, which return the current call stack
}
