% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{callstack_message}
\alias{callstack_message}
\title{Create callstack message from callstack data}
\usage{
callstack_message(message, stack)
}
\arguments{
\item{message}{Header mesasge}

\item{stack}{callstack data}
}
\value{
message with callstack data
}
\description{
Create callstack message which include the information of callstack data.
}
