% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{callstack_trace}
\alias{callstack_trace}
\title{Create callstack data at last error}
\usage{
callstack_trace(IgnoreLevel = 0)
}
\arguments{
\item{IgnoreLevel}{Number of callstack data level which should be ignored from return value}
}
\value{
callstack dataframe with name, line and filename.
}
\description{
Create callstack data from .traceback function, which return the call stack at the last error
}
