% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{catch_either}
\alias{catch_either}
\title{Try excute function which might throw exception.}
\usage{
catch_either(value, alt_value, catchfn = NULL)
}
\arguments{
\item{value}{Value which might be exception.}

\item{alt_value}{return value when expr throw exception.}

\item{catchfn}{Function which will be called when expr throw exception with error message as argument.}
}
\value{
return value if expr don't throw, or exception if it throw.
}
\description{
Try execute function which might throw exception.
}
