% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cold.colors}
\alias{cold.colors}
\title{Create cold color set.}
\usage{
cold.colors(n, val.white = 0.2, rate.white = 0.4)
}
\arguments{
\item{n}{Size of color set.}

\item{val.white}{Strength of white color.}

\item{rate.white}{Changing rate of the white color}
}
\value{
Created color set.
}
\description{
Create cold-color based color set.
}
