% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{count}
\alias{count}
\title{Return counted number which is matched to the argument}
\usage{
count(what, from, condition = `==`)
}
\arguments{
\item{what}{value of sequence for finding}

\item{from}{index of this argument is returned}

\item{condition}{condition for finding. Default is "=="}
}
\value{
List of count number.
}
\description{
Count what in from
}
\examples{
count(c(1,3,5), c(0,1,2,3,3))
# c(1,2,0)
}
