% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{count_row}
\alias{count_row}
\title{Return counted number which is matched to the any argument in each row.}
\usage{
count_row(what, fromMD, condition = `==`)
}
\arguments{
\item{what}{value of sequence for finding}

\item{fromMD}{index of this argument is returned}

\item{condition}{condition for finding. Default is "=="}
}
\value{
List of count number.
}
\description{
Count what in from
}
\examples{
count_row(c(1,3,5), matrix(c(0,0,1,2,0,0,3,0,0,3,4,4),ncol=2,byrow=TRUE))
# c(1,2,0)
}
