% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cud.colorset}
\alias{cud.colorset}
\title{CUD v3 color palet as the color universal design}
\usage{
cud.colorset()
}
\value{
List of colors
}
\description{
CUD v3 is the color palet for the color universal design, i.e., considring the accessibility for color blinders.
}
