% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{cud.pallete}
\alias{cud.pallete}
\title{CUD v3 color palette as the color universal design}
\usage{
cud.pallete(n = 10)
}
\arguments{
\item{n}{Size of color set.}
}
\value{
Created color set.
}
\description{
CUD v3 is the color palet for the color universal design, i.e., considring the accessibility for color blinders.
}
