% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{date_to_weekno}
\alias{date_to_weekno}
\title{Calculate weekno from Date}
\usage{
date_to_weekno(Date)
}
\arguments{
\item{Date}{the focal date}
}
\value{
Week number
}
\description{
Calculate weekno from Date
}
