% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{datestr_format_full.fullcode}
\alias{datestr_format_full.fullcode}
\title{formatting date string}
\usage{
datestr_format_full.fullcode(
  datestr,
  Date.beg = NULL,
  Date.end = NULL,
  strict = TRUE,
  use.range = TRUE,
  use.jpera = TRUE,
  use.exceldate = FALSE,
  exceldate.origin = as.Date("1904-01-01")
)
}
\arguments{
\item{datestr}{target character or sequence of character.}

\item{Date.beg}{potentially begin date; used for validation and estimation.}

\item{Date.end}{potentially end date; used for validation and estimation.}

\item{strict}{return NA if read datestr is out of Date.beg-Date.end range}

\item{use.range}{Allow range date format}

\item{use.jpera}{Allow Japanese era case (e.g., R2.4.5 or 令和3年4月11日).}

\item{use.exceldate}{Allow to check the value which is originally excel date but occationally transformed to normal value.}

\item{exceldate.origin}{Origin of excel date. In default, it is 1900-01-01 if the file is originally created on Windows and 1904-01-01 on Mac.}
}
\value{
transformed date string like "2021.03.04-2023.05.31"
}
\description{
unify valid format of date
}
