% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{distlim_hclust_clustering}
\alias{distlim_hclust_clustering}
\title{Categorize multi dimension data based on the relative distance vs cluster size using hclust function}
\usage{
distlim_hclust_clustering(data, thr = 5, kmax = 10, method = "ward.D2")
}
\arguments{
\item{data}{Matrix data for clustering. Row is element and Col is axis.}

\item{thr}{Threshold value of the relative distance/size of clusters.}

\item{kmax}{Max cluster number.}

\item{method}{Clustering method in hclust.}
}
\value{
List with size: number of cluster and group: #cluster of each element.
}
\description{
Based on the h-clustering algorithm, find optimal cluster num by keeping multiple cluster are not overlapped.
}
