% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{either}
\alias{either}
\title{Return y if x satisfies given condition, otherwise return x}
\usage{
either(x, y, condition = is.na)
}
\arguments{
\item{x}{target value}

\item{y}{alternatively returned value}

\item{condition}{a function return logical values for each x}
}
\value{
x or y, whose length is euqal to length_x
}
\description{
Return y if x satisfies given condition, otherwise return x
}
