% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{ensure}
\alias{ensure}
\title{Return y if x does not exist, otherwise return x}
\usage{
ensure(x, y)
}
\arguments{
\item{x}{target value}

\item{y}{alternatively returned value}
}
\value{
x or y
}
\description{
Return y if x does not exist, otherwise return x
}
