% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{error_points}
\alias{error_points}
\title{Add points with error bars}
\usage{
error_points(x, y, upper, lower = NULL, head.length = 0.05,
  horizontal = FALSE, ...)
}
\arguments{
\item{x}{vector of x-axis.}

\item{y}{vector of y-axis.}

\item{upper}{Error value of the upper boundary.}

\item{lower}{Error value of the lower boundary. In default, lower = -upper.}

\item{head.length}{Length of the head of error bar.}

\item{horizontal}{Logical: if true the error bar for x will be drawn.}

\item{...}{arguments for arrows and points function}
}
\description{
Drow points with error bars.
}
