% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{escape_md}
\alias{escape_md}
\title{Escape all markdown character.}
\usage{
escape_md(str)
}
\arguments{
\item{str}{character vector}
}
\value{
Returns a character vector.
}
\description{
Escape all markdown character.
}
