% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape all regex character.}
\usage{
escape_regex(str)
}
\arguments{
\item{str}{character vector}
}
\value{
Returns a character vector.
}
\description{
Escape all regex character.
}
