% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{execute}
\alias{execute}
\title{Execute given exe file with arguments.}
\usage{
execute(exe, ...)
}
\arguments{
\item{exe}{Path of exe file.}

\item{...}{optional arguments for exe file.}
}
\description{
Execute given exe file with arguments.
}
