% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.abspath}
\alias{file.abspath}
\title{Return absolute path ==BUG EXIST REPLACE BY file.stdpath==}
\usage{
file.abspath(filepath, filedir, strict = TRUE)
}
\arguments{
\item{filepath}{file path (character)}

\item{filedir}{file directory}

\item{strict}{logical: TRUE ~ remove unnecessary reference, e.g., "a/b/../c".}
}
\value{
absolute file path
}
\description{
Return logical value which represents whether the given path is absolute path or not.
}
