% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.backup}
\alias{file.backup}
\title{Backup given file}
\usage{
file.backup(
  filepath,
  backup = "bak",
  backup.dir = NULL,
  timeformat = "\%y\%m\%d_\%H\%M\%S_"
)
}
\arguments{
\item{filepath}{file path (character)}

\item{backup}{backup directory name; ignored if backup.dir is defined.}

\item{backup.dir}{backup directory path; default is the same directory of given file.}

\item{timeformat}{set backup head name.}
}
\value{
filepath
}
\description{
Make backup of the given file in backup directory
}
