% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.dir}
\alias{file.dir}
\title{Get directory of file}
\usage{
file.dir(filepath)
}
\arguments{
\item{filepath}{file path (character)}
}
\value{
directory of file of given file path.
}
\description{
Get directory of file from file path.
}
