% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.ext}
\alias{file.ext}
\title{Get file extension}
\usage{
file.ext(filepath)
}
\arguments{
\item{filepath}{file path (character)}
}
\value{
extension of given file path.
}
\description{
Get file extension from file path.
}
