% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.fout}
\alias{file.fout}
\title{generate file output}
\usage{
file.fout(filepath, open = "w", encoding = getOption("encoding"))
}
\arguments{
\item{filepath}{file path (character)}

\item{open}{open mode}

\item{encoding}{encoding}
}
\value{
fout function
}
\description{
Generate file output
}
