% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.is_abspath}
\alias{file.is_abspath}
\title{Check whether the given path is absolute path or not.}
\usage{
file.is_abspath(filepath)
}
\arguments{
\item{filepath}{file path (character)}
}
\value{
logical: true if it is absolute path.
}
\description{
Return logical value which represents whether the given path is absolute path or not.
}
