% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.name}
\alias{file.name}
\title{Get file name without extension}
\usage{
file.name(filepath)
}
\arguments{
\item{filepath}{file path (character)}
}
\value{
file name of given file path.
}
\description{
Get file name without extension from file path.
}
