% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.remove_dir}
\alias{file.remove_dir}
\title{Get file name with extension}
\usage{
file.remove_dir(filepath)
}
\arguments{
\item{filepath}{file path (character)}
}
\value{
file name with extension of given file path.
}
\description{
Get file name with extension from file path.
}
