% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.remove_ext}
\alias{file.remove_ext}
\title{Get dir and name without extension}
\usage{
file.remove_ext(filepath)
}
\arguments{
\item{filepath}{file path (character)}
}
\value{
extension of given file path.
}
\description{
Remove ext part from the path
}
