% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.replace_ext}
\alias{file.replace_ext}
\title{Replace without extension}
\usage{
file.replace_ext(filepath, ext)
}
\arguments{
\item{filepath}{file path (character)}

\item{ext}{new extension}
}
\value{
extension of given file path.
}
\description{
Remove ext part from the path
}
