% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.stdpath}
\alias{file.stdpath}
\title{Return standardized path}
\usage{
file.stdpath(path, rootpath = NA_character_, strict = TRUE, stdroot = FALSE)
}
\arguments{
\item{path}{target file/dir path (character)}

\item{rootpath}{root file/dir path of the target path. if path is absolute path, rootpath is ignored}

\item{strict}{logical: TRUE ~ remove unnecessary reference "..", e.g., "a/b/../c".}

\item{stdroot}{logical: TRUE ~ skip rootpath standarization}
}
\value{
standardized path
}
\description{
Return standardized path from rootpath. head/tail spaces, current directory
}
