% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{find_first_row}
\alias{find_first_row}
\title{Return list of first row in fromMD in any of which is matched to what.}
\usage{
find_first_row(what, fromMD, condition = `==`)
}
\arguments{
\item{what}{value of sequence for finding}

\item{fromMD}{multi-dimensional target.}

\item{condition}{condition for finding. Default is "=="}
}
\value{
List of first row which is found in from at each what. NA if it is not found.
}
\description{
Return list of row in fromMD in any of which is matched to what.
}
\examples{
find_first_row(c(1,3,5), matrix(c(0,0,1,2,0,0,3,0,0,3,4,4),ncol=2,byrow=TRUE))
#' # c(2,4,NA)
}
