% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{find_unique}
\alias{find_unique}
\title{Return index of from which is uniquely matched to the argument}
\usage{
find_unique(what, from, condition = `==`)
}
\arguments{
\item{what}{value of sequence for finding}

\item{from}{index of this argument is returned}

\item{condition}{condition for finding. Default is "=="}
}
\value{
List of index which is found in from. NA if it is not found, or multiple index are found.
}
\description{
Find unique index number in from at which is equal to what.
}
\examples{
find_unique(c(1,3,5), c(0,1,2,3,3))
# c(2,NA,NA)
}
