% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{formatted_output}
\alias{formatted_output}
\title{create output function list.}
\usage{
formatted_output(
  outpath = NULL,
  errpath = NULL,
  head_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  use_utc = FALSE,
  use_pid = FALSE
)
}
\arguments{
\item{outpath}{Optional; output file path which log all of output.}

\item{errpath}{Optional; error file path which log only warning and error output.}

\item{head_format}{header format of each row. Default is \%Y-\%m-\%d \%H:\%M:\%S.}

\item{use_utc}{Logical; if TRUE, the head format is filled by UTC and head end with 'Z'.}

\item{use_pid}{Logical; if TRUE, the pid is added as '@xxxxx'.}
}
\value{
print, warning and error functions.
}
\description{
crate list of functions including print, warning and error, each of which accept format style like sprintf.
}
