% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{get_cross_point}
\alias{get_cross_point}
\title{Get cross point between two segments}
\usage{
get_cross_point(
  seg1.a_x,
  seg1.a_y,
  seg1.b_x,
  seg1.b_y,
  seg2.a_x,
  seg2.a_y,
  seg2.b_x,
  seg2.b_y
)
}
\arguments{
\item{seg1.a_x}{x value of point A on segment 1}

\item{seg1.a_y}{y value of point A on segment 1}

\item{seg1.b_x}{x value of point B on segment 1}

\item{seg1.b_y}{y value of point B on segment 1}

\item{seg2.a_x}{x value of point A on segment 2}

\item{seg2.a_y}{y value of point A on segment 2}

\item{seg2.b_x}{x value of point B on segment 2}

\item{seg2.b_y}{y value of point B on segment 2}
}
\value{
vector whose first and second elements are x and y.
}
\description{
Check if two segments has thier cross point or not.
}
\note{
This function return incorrect values when given segments has no cross point.
}
