% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{green.colors}
\alias{green.colors}
\title{Create green color set.}
\usage{
green.colors(n, val.white = 0.2, rate.white = 0.5, rate.dark = 0.5)
}
\arguments{
\item{n}{Size of color set.}

\item{val.white}{Strength of white color.}

\item{rate.white}{Changing rate for bright side.}

\item{rate.dark}{Changing rate for dark side.}
}
\value{
Created color set.
}
\description{
Create green-color based color set.
}
