% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{hankana2zenkana}
\alias{hankana2zenkana}
\title{Replace hankaku-kana by zenkaku-kana}
\usage{
hankana2zenkana(x)
}
\arguments{
\item{x}{target character}
}
\value{
replaced character
}
\description{
Replace hankaku-kana by zenkaku-kana.
}
