% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{has_cross_point}
\alias{has_cross_point}
\title{Check if the segment seg1 is cross with seg2}
\usage{
has_cross_point(
  seg1.a_x,
  seg1.a_y,
  seg1.b_x,
  seg1.b_y,
  seg2.a_x,
  seg2.a_y,
  seg2.b_x,
  seg2.b_y
)
}
\arguments{
\item{seg1.a_x}{x value of point A on segment 1}

\item{seg1.a_y}{y value of point A on segment 1}

\item{seg1.b_x}{x value of point B on segment 1}

\item{seg1.b_y}{y value of point B on segment 1}

\item{seg2.a_x}{x value of point A on segment 2}

\item{seg2.a_y}{y value of point A on segment 2}

\item{seg2.b_x}{x value of point B on segment 2}

\item{seg2.b_y}{y value of point B on segment 2}
}
\value{
logical value: TRUE when segment1 and segment2 has a cross point.
}
\description{
Check if two segments has thier cross point or not.
}
