% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histgram.R
\name{hist.smoothing}
\alias{hist.smoothing}
\title{Smoothing histgram sequence by using simple moving average method}
\usage{
\method{hist}{smoothing}(hist.data, width)
}
\arguments{
\item{hist.data}{Histgram sequence}

\item{width}{Width of the simple moving average method. width=1 means no smoothing. This value should be an intergral value.}
}
\value{
Smoothing histgram sequence
}
\description{
Smoothing histgram sequence by using simple moving average method
}
\examples{
hist.data = c(1,5,7,5,5,4,3,1,0,0,0,4,4,8,5,5,5,5,3,3,3,7,7,8,3,3,1,3,5,0,0,2,3,3,0)
smooth = hist.smoothing(hist.data,3)
plot(hist.data,type="b")
lines(1:length(hist.data),smooth,col="red",lwd=2)
}
