% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{image_polygon}
\alias{image_polygon}
\title{Add polygon on image}
\usage{
image_polygon(x, y, z, ...)
}
\arguments{
\item{x}{vector of x-axis}

\item{y}{vector of y-axis}

\item{z}{matrix where polygon is drawn if TRUE}

\item{...}{arguments for polygon function}
}
\description{
Drow polygon on image where z==true
}
