% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{is_exception}
\alias{is_exception}
\title{Check whether the argument is exception}
\usage{
is_exception(arg)
}
\arguments{
\item{arg}{target argument.}
}
\value{
Logical; TRUE:exception, FALSE: otherwise.
}
\description{
Check whether the argument is exception
}
