% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{is_on_linux}
\alias{is_on_linux}
\title{Check if the program run on the Linux-base or windows from path system.}
\usage{
is_on_linux()
}
\value{
Logical values: TRUE on Linux or Mac, FALSE on Windows.
}
\description{
This function return TRUE when the program run on the linux environment.
}
