% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{jpera_to_date}
\alias{jpera_to_date}
\title{get Date type from Japanese Era name}
\usage{
jpera_to_date(str, tryFormats = c("\%E-\%m-\%d", "\%E.\%m.\%d"), ...)
}
\arguments{
\item{str}{target character or sequence of character.}

\item{tryFormats}{format of str for reading JP era and other date data. \%E mean era. The other options follows as.Date function.}
}
\value{
Date class
}
\description{
Work like as.Date but Japanese Era can be allowed, which is represented as \%E in format. Default formats are "R1-04-03" or "R1.04.03".
}
