% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{jpera_to_y2str}
\alias{jpera_to_y2str}
\title{transform Japanese-Era string into AD-year string with two digit.}
\usage{
jpera_to_y2str(str)
}
\arguments{
\item{str}{target character or sequence of character.}
}
\value{
transformed str
}
\description{
transform Japanese-Era string into AD-year string with two digit (i.e., R1 -> 19)
}
