% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{jpera_to_ystr}
\alias{jpera_to_ystr}
\title{transform Japanese-Era string into AD-year string.}
\usage{
jpera_to_ystr(str)
}
\arguments{
\item{str}{target character or sequence of character.}
}
\value{
transformed str
}
\description{
transform Japanese-Era string into AD-year string (i.e., R1 -> 2019)
}
