% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{match.order}
\alias{match.order}
\title{Return index numbers of the target sequence which is matching to the obj,}
\usage{
match.order(obj, trg)
}
\arguments{
\item{obj}{Value, vector or matrix}

\item{trg}{Target sequence}
}
\value{
Found once: Index numbers, Found multiple times: NA, Didn't find: 0
}
\description{
Find index number at which trg is equal to obj. Return 0 if it is not found in trg, and return NA if it has been multitmes found.
}
\examples{
x = c(115,103,99,120)
trg = c(1:50 + 100,120)
match.order(x,trg)
# c(15,3,0,NA)
}
