% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{match.order.reduce}
\alias{match.order.reduce}
\title{Return reduced value by the given target sequence.}
\usage{
match.order.reduce(from, fromval, trg, ini = 0, func = sum)
}
\arguments{
\item{from}{Value, vector or matrix}

\item{fromval}{Value, vector or matrix}

\item{trg}{Finding target sequence}

\item{ini}{initial values of trg}

\item{func}{function for reduce (default=sum)}
}
\description{
Find values at which trg is equal to from.
}
