% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{matlab.pallete}
\alias{matlab.pallete}
\title{matlab color palette as the color universal design}
\usage{
matlab.pallete(n = 10)
}
\arguments{
\item{n}{Size of color set.}
}
\value{
Created color set.
}
\description{
Return color palette used in matlab.
}
