% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{matrix_from_func.nonseqarg}
\alias{matrix_from_func.nonseqarg}
\title{Create matrix from function without vector argument support.}
\usage{
matrix_from_func.nonseqarg(x, y, f, ...)
}
\arguments{
\item{x}{First argument vectors, and length of x is the row of created matrix.}

\item{y}{Second argument vectors, and length of y is the col of created matrix.}

\item{f}{Function for creating matrix.}

\item{...}{Arguments are passed to function f.}
}
\value{
matrix whose elements are obtained from f with possible all combination of x and y
#' @examples
x=0:3
y=0:2
matrix_from_func.nonseqarg(x,y,function(x,y){return(x+y)})
#((0,1,2,3),(1,2,3,4),(2,3,4,5)
}
\description{
This function create the matrix from a defined function and two sequences. The used function don't have to support the sequence-style return from sequence-style arguments. Elements of the created matrix are obtained from the function with two arguments, which are possible combinations of elements from two sequences.
}
