% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{matrix_from_func}
\alias{matrix_from_func}
\title{Create matrix from function}
\usage{
matrix_from_func(x, y, f, ...)
}
\arguments{
\item{x}{First argument vectors, and length of x is the row of created matrix.}

\item{y}{Second argument vectors, and length of y is the col of created matrix.}

\item{f}{Function for creating matrix. This function should support the sequence-style return from sequence-style arguments.}

\item{...}{Arguments are passed to function f.}
}
\value{
matrix whose elements are obtained from f with possible all combination of x and y
}
\description{
This function create the matrix from a defined function and two sequences. The used function should support the sequence-style return from sequence-style arguments. Elements of the created matrix are obtained from the function with two arguments, which are possible combinations of elements from two sequences.
}
\examples{
x=0:3
y=0:2
fn = function(x,y){return(x+y)}
matrix_from_func(x,y,fn)
#((0,1,2,3),(1,2,3,4),(2,3,4,5))
}
