% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{matrix_from_function}
\alias{matrix_from_function}
\title{matrix_from_function}
\usage{
matrix_from_function(f, x, y, ...)
}
\arguments{
\item{f}{using function with two arguments}

\item{x}{first argument vectors}

\item{y}{second argument vectors}
}
\value{
matrix whose elements are obtained from f with possible all combination of x and y
}
\description{
Create matrix from function
}
\examples{
x=0:3
y=0:2
matrix_from_function(function(x,y){return(x+y)},x,y)
#((0,1,2,3),(1,2,3,4),(2,3,4,5)
}

