% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{matrix_na}
\alias{matrix_na}
\title{Create matrix with NA}
\usage{
matrix_na(nx, ny)
}
\arguments{
\item{nx}{Length of row.}

\item{ny}{Length of col.}
}
\value{
Matrix whose all elements are NA
}
\description{
Create matrix with NA
}
\examples{
matrix_na(3,2)
#((NA,NA),(NA,NA),(NA,NA))
}
