% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{mids}
\alias{mids}
\title{Return vector of the mid points of the argument}
\usage{
mids(x)
}
\arguments{
\item{x}{terget vector}
}
\value{
vector of mid points
}
\description{
Create the sequence of the mid points of vector, i.e., (x[-1]+x[-length(x)])/2
}
\examples{
x = c(0.0,0.2,0.6,1.2)
ans = mids(x)
# ans == c(0.1,0.4,0.9)
}
