% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histgram.R
\name{minmaxhist_to_hist}
\alias{minmaxhist_to_hist}
\title{Translate histgram matrix with min values and max values to a unified histgram matrix (i.e., same breaks)}
\usage{
minmaxhist_to_hist(minseq, maxseq, histmx, minval = NA, maxval = NA,
  len = NA)
}
\arguments{
\item{minseq}{min values of source histgram matrix.}

\item{maxseq}{max values of source histgram matrix.}

\item{histmx}{source histgram matrix.}

\item{minval}{min value of a unified histrgam.}

\item{maxval}{max value of a unified histrgam.}

\item{len}{length of a unified histrgam.}
}
\value{
List of new unified histgram matrix (hist), breaks (break) and mid values of each element (mids)
}
\description{
This function translate the histgram set with different min/max values into a histgram with a common breaks.
}
