% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{new_env}
\alias{new_env}
\title{generate new environment with arglist.}
\usage{
new_env(env, arglist = NULL)
}
\arguments{
\item{env}{parent environment}

\item{arglist}{argument list; argument without name is ignored.}
}
\value{
environment
}
\description{
generate new environment from parent with arglist which are assigned with those names.
}
