% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{newton_planeroot}
\alias{newton_planeroot}
\title{Find root points on x-y plane by using newton method}
\usage{
newton_planeroot(func, xlim = c(0, 1), ylim = c(0, 1), n = 1001, ...)
}
\arguments{
\item{func}{Function  whose root points are searched}

\item{xlim}{min and max value of x-axis.}

\item{ylim}{min and max value of y-axis.}

\item{n}{step num for small_step search}

\item{...}{optional arguments to be passed to "func"}
}
\value{
list(x,y). x and y are sequence of points.
}
\description{
Return vectors which are sequence of root points on the given grid (x,y).
}
