% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{newton_root}
\alias{newton_root}
\title{Find root of a function using combination of small_step method and newton method.}
\usage{
newton_root(func, step, ...)
}
\arguments{
\item{func}{the function for which the root is sought.}

\item{step}{a vector containing the step values}

\item{...}{optional arguments to be passed to "func"}
}
\value{
a vector which containing root values
}
\description{
Returns roots of a focal function by checking the sign change in each step.
}
\examples{
f = function(x,y){return(x*x-y*x)}
ans = newton_root(f, seq(-2,2,length=1001),y=1.333)
#ans == c(0.000,1.333)
}
