% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{os_notification}
\alias{os_notification}
\title{Throw toast as the OS notification}
\usage{
os_notification(from, message)
}
\arguments{
\item{from}{Notifying app name}

\item{message}{Notifying message}
}
\value{
system call result.
}
\description{
Output characters as toast. Available only on Windows.
}
