% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{pd.mean}
\alias{pd.mean}
\title{Return mean value of the probability distribution}
\usage{
pd.mean(x, pd)
}
\arguments{
\item{x}{axis value}

\item{pd}{probability distribution or histgram data}
}
\value{
mean value of the given pd
}
\description{
Calculate the mean value of the probability distribution or histgram data.
}
