% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{pd.median}
\alias{pd.median}
\title{Return median of the probability distribution}
\usage{
pd.median(x, pd)
}
\arguments{
\item{x}{axis value}

\item{pd}{probability distribution or histgram data}
}
\value{
median of the given pd
}
\description{
Calculate the median of the probability distribution or histgram data.
}
