% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{pd.skewness}
\alias{pd.skewness}
\title{Return skewness of the probability distribution}
\usage{
pd.skewness(x, pd)
}
\arguments{
\item{x}{axis value}

\item{pd}{probability distribution or histgram data}
}
\value{
skewness of the given pd
}
\description{
Calculate the skewness of the probability distribution or histgram data.
}
