% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{pd.var}
\alias{pd.var}
\title{Return variance of the probability distribution}
\usage{
pd.var(x, pd)
}
\arguments{
\item{x}{axis value}

\item{pd}{probability distribution or histgram data}
}
\value{
variance of the given pd
}
\description{
Calculate the variance of the probability distribution or histgram data.
}
