% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.null}
\alias{plot.null}
\title{Draw plot without any data}
\usage{
\method{plot}{null}(...)
}
\arguments{
\item{...}{Arguments are passed to function plot.}
}
\description{
This function just call plot function with parameter type = "n".
}
