% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{rand_char}
\alias{rand_char}
\title{Generate random character sequences}
\usage{
rand_char(n, len, lower_case = TRUE, upper_case = TRUE, number = TRUE)
}
\arguments{
\item{n}{number of str}

\item{len}{length of each str}

\item{lower_case}{logical: include lower case alphabets}

\item{upper_case}{logical: include upper case alphabets}

\item{number}{logical: include numbers}
}
\value{
sequences of string.
}
\description{
return n strings with length len.
}
