% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rb.colors}
\alias{rb.colors}
\title{Create red-blue color set.}
\usage{
rb.colors(n, low = 0.25, high = 0.9)
}
\arguments{
\item{n}{Size of color set.}

\item{low}{The hue at the sides of the color set.}

\item{high}{The hue at the center of the color set.}
}
\value{
Created color set.
}
\description{
Create red-blue based color set.
}
