% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{reduce_cols}
\alias{reduce_cols}
\title{reduce multiple columns into single column.}
\usage{
reduce_cols(
  data,
  col,
  ...,
  .keep = FALSE,
  .before = NULL,
  .after = NULL,
  .reduce = function(x, y) {
     return(x)
 }
)
}
\arguments{
\item{data}{table}

\item{col}{column}

\item{...}{merged columns}

\item{.keep}{whether the merged column are kept or removed.}

\item{.before}{location of column before brabra.}

\item{.after}{location of column after brabra.}

\item{.reduce}{reduce function; default is the first column which is not NA}
}
\value{
data with the reseved columns
}
\description{
reduce multiple columns into single column by funton .reduce.
}
