% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reload_package}
\alias{reload_package}
\title{reload all packages which have been loaded in any environment.}
\usage{
reload_package(env = parent.env())
}
\arguments{
\item{env}{target environment}
}
\description{
reload all packages which have been loaded in any environment.
}
