% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{remove.na}
\alias{remove.na}
\title{Remove na elements from sequences.}
\usage{
remove.na(seq)
}
\arguments{
\item{seq}{Terget sequence.}
}
\value{
Sequence including only non.na elements.
}
\description{
Remove na elements from sequences.
}
\examples{
x=c(0,NA,2,3,NA,5,NA)
remove.na(x)
#(0,2,3,5)
}
