% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{replace_by}
\alias{replace_by}
\title{Replace elements by other values}
\usage{
replace_by(target, from, to, other = NULL)
}
\arguments{
\item{target}{target vector, matrix or other types}

\item{from}{the element values which should be replaced}

\item{to}{the values which should be set for the replaced elements}

\item{other}{if it is not NULL value, the target elements which is not hit to any of from values are replaced by other.}
}
\value{
replaced results
}
\description{
Replace elements by other values
}
