% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{replace_by_threshold}
\alias{replace_by_threshold}
\title{Select elements depending on thresholds}
\usage{
replace_by_threshold(target, thresholds, to)
}
\arguments{
\item{target}{target vector, matrix or other types}

\item{thresholds}{the threshold values (should be sorted)}

\item{to}{the values which should be set depending on the thresholds}
}
\value{
replaced results
}
\description{
Select elements depending on thresholds
}
