% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{reserve_cols}
\alias{reserve_cols}
\title{reserve columns with the given names in the table}
\usage{
reserve_cols(data, col_names)
}
\arguments{
\item{data}{table}

\item{col_names}{column names for reserve}
}
\value{
data with the reserved columns
}
\description{
the column is generated when there are not.
}
