% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{run_as_filedrop_script}
\alias{run_as_filedrop_script}
\title{Run as Rscript with dragged files}
\usage{
run_as_filedrop_script(fn, ...)
}
\arguments{
\item{fn}{Function whose first argument is a drag&drop file}

\item{...}{Arguments are passed to function f.}
}
\description{
Run given function as Rscript with drag&drop files.
}
