% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{run_as_rscript}
\alias{run_as_rscript}
\title{Run as Rscript with dragged files}
\usage{
run_as_rscript(fn, ...)
}
\arguments{
\item{fn}{Funtion whose first argument is a dragged file}

\item{...}{Arguments are passed to function f.}
}
\description{
Run given function as Rscript with dragged files.
}
