% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{run_rscript}
\alias{run_rscript}
\title{Execute given Rscript file with arguments.}
\usage{
run_rscript(rfile, ...)
}
\arguments{
\item{rfile}{Path of Rscript file}

\item{...}{optional arguments for rscript file.}
}
\description{
Execute given Rscript file with arguments.
}
