% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{save_plot}
\alias{save_plot}
\title{Save plot expression into laster/vector files}
\usage{
save_plot(plotexpr, filename, width = 4, height = 3, ppi = 300, ...)
}
\arguments{
\item{plotexpr}{plot function or plot expression, which is evaluated "lazily".}

\item{filename}{File name. file extension determines the file format.}

\item{width}{width of figure in inch.}

\item{height}{height of figure in inch.}

\item{ppi}{pixel per inchi; used only for raster format.}

\item{...}{arguments for par, e.g., cex or mex.}
}
\description{
Save plot into a figure file with given extension
}
