% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{search_cross_point}
\alias{search_cross_point}
\title{Find equal points from point groups.}
\usage{
search_cross_point(xseq1, yseq1, xseq2, yseq2)
}
\arguments{
\item{xseq1}{x-values of first point group}

\item{yseq1}{y-values of first point group}

\item{xseq2}{x-values of second point group}

\item{yseq2}{y-values of second point group}
}
\value{
list(x,y). x and y are sequence of points.
}
\description{
Return equal points from two point groups, seq1 and seq2.
}
