% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{seq_na}
\alias{seq_na}
\title{Create sequence with NA}
\usage{
seq_na(num)
}
\arguments{
\item{num}{Length of sequence.}
}
\value{
Sequence whose all elements are NA
}
\description{
Create sequence with NA
}
\examples{
seq_na(4)
#(NA,NA,NA,NA)
}
