% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sizedist_clust}
\alias{sizedist_clust}
\title{Categorize multi dimension data based on the relative distance vs cluster size}
\usage{
sizedist_clust(data, rate = 2, iter = 10, kmax = 10)
}
\arguments{
\item{data}{Matrix data for clustering. Row is element and Col is axis.}

\item{rate}{Threshold rate of clustersize / distance}

\item{iter}{Number of iteration for kmeans.}

\item{kmax}{Max cluster number.}
}
\value{
List with size: number of cluster and cluster: #cluster of each element.
}
\description{
Based on the kmeans clustering algorithm, find optimal cluster num by keeping multiple cluster are not overlapped.
}
