% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{small_step_planeroot_from_matrix}
\alias{small_step_planeroot_from_matrix}
\title{Find root points on x-y plane from logical matrix and x,y axes.}
\usage{
small_step_planeroot_from_matrix(x, y, z)
}
\arguments{
\item{x}{sequence of x-axis}

\item{y}{sequence oy y-axis}

\item{z}{logical matrix of function values whose root are searched}
}
\value{
list(x,y). x and y are sequence of points.
}
\description{
Return vectors which are sequence of root points on the given grid (x,y).
}
\note{
nrow(z) and ncol(z) should be equal to length(x) and length(y), respectively.
}
