% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{small_step_root}
\alias{small_step_root}
\title{Find root of a function using small_step method}
\usage{
small_step_root(func, step, ...)
}
\arguments{
\item{func}{the function for which the root is sought.}

\item{step}{a vector containing the step values}

\item{...}{optional arguments to be passed to "func"}
}
\value{
data.frame with two col, "lower" and "upper" which are the range of the solutions. root
}
\description{
This function returns roots of a focal function by checking the sign change in each step.
}
\examples{
f = function(x,y){return(x*x-y*x)}
ans = small_step_root(f, seq(-2,2,length=41),y=1.333)
# ans == data.frame("lower"=c(0.0,1.3),"upper"=c(0.0,1.4))
}
