% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{solve_numeric}
\alias{solve_numeric}
\title{solve stable fraction numerically}
\usage{
solve_numeric(mx, hint = NULL, tol = 1e-12, inidf = 0.01)
}
\arguments{
\item{mx}{target matrix for solution}

\item{hint}{vector which potentially close to the solution (for initial value)}

\item{tol}{tolerance of error}

\item{inidf}{time step}
}
\value{
numeric sequence of result vector.
}
\description{
Return vectors which are solution of matrix.
}
