% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp.R
\name{source_cpp}
\alias{source_cpp}
\title{source C++ code from a file with include/library path.}
\usage{
source_cpp(file, I = NULL, L = NULL, l_opt = NULL, ...)
}
\arguments{
\item{file}{A character string giving the path name of a file}

\item{I}{A character (or vector of character) giving the path name of include directory}

\item{L}{A character (or vector of character) giving the path name of library directory}

\item{l_opt}{A character (or vector of character) giving the additional library option (e.g. LibOptions = c("-lm"," -lmpc"))}

\item{...}{other arguments for sourceCpp}
}
\value{
a vector which containing root values
}
\description{
call Rcpp::sourceCpp with temporally defining include/library path
}
