% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{start_ownself}
\alias{start_ownself}
\title{Execute ownself in parallel with arguments.}
\usage{
start_ownself(ownpath, ...)
}
\arguments{
\item{ownpath}{Own file path}

\item{...}{optional arguments for Rscript file.}
}
\description{
Run ownself as Rscript in parallel with arguments. In parallel started files, this function does not start new process.
}
