% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{start_rscript}
\alias{start_rscript}
\title{Execute given Rscript file in parallele with arguments.}
\usage{
start_rscript(rfile, ...)
}
\arguments{
\item{rfile}{Path of Rscript file}

\item{...}{optional arguments for rscript file.}
}
\description{
Execute given Rscript file in parallele with arguments.
}
