% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_either}
\alias{str_either}
\title{Return y if x is NA or empty character, otherwise return x}
\usage{
str_either(x, y)
}
\arguments{
\item{x}{target value}

\item{y}{alternatively returned value}
}
\value{
x or y, whose length is euqal to length_x
}
\description{
Return y if x is NA, otherwise return x
}
