% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_first_appear}
\alias{str_first_appear}
\title{Wherher the given string appear first time in the given string vector}
\usage{
str_first_appear(string, similarity = 3)
}
\arguments{
\item{string}{target character vector.}

\item{similarity}{threshold similarity}
}
\value{
logical: TRUE if the character element appear first time.
}
\description{
str can be uniqued by using string[str_first_appear(string)].
}
