% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_kana_to_zen}
\alias{str_kana_to_zen}
\title{Replace hankaku-kana by zenkaku-kana}
\usage{
str_kana_to_zen(x)
}
\arguments{
\item{x}{target character}
}
\value{
replaced character
}
\description{
Replace hankaku-kana by zenkaku-kana.
}
