% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_number_to_han}
\alias{str_number_to_han}
\title{Replace zenkaku-number by hankaku-number}
\usage{
str_number_to_han(x)
}
\arguments{
\item{x}{target character}
}
\value{
replaced character
}
\description{
Replace zenkaku-number by hankaku-number
}
