% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_remove_dakuten}
\alias{str_remove_dakuten}
\title{Remove daku-on, han-daku-on}
\usage{
str_remove_dakuten(x)
}
\arguments{
\item{x}{target character}
}
\value{
replaced character
}
\description{
Remove daku-on, han-daku-on
}
