% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_similar}
\alias{str_similar}
\title{Check similarity of given string with target.}
\usage{
str_similar(string, target, similarity = 3)
}
\arguments{
\item{string}{character for checking similarity}

\item{target}{character for compared target}

\item{similarity}{threshold similarity}
}
\value{
logical: TRUE if string is enough similar with target.
}
\description{
Check similarity of given string with target.
}
