% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_similar_pair}
\alias{str_similar_pair}
\title{Check similarity pair of given string with themselves.}
\usage{
str_similar_pair(string, similarity = 3, only_sub = FALSE)
}
\arguments{
\item{string}{character for checking similarity}

\item{similarity}{threshold similarity}

\item{only_sub}{count only substitute}
}
\value{
logical: TRUE if string is enough similar with target.
}
\description{
Check similarity pair of given string with themselves.
}
