% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_symbol_to_han}
\alias{str_symbol_to_han}
\title{Replace zenkaku-symbol by hankaku-symbol}
\usage{
str_symbol_to_han(x)
}
\arguments{
\item{x}{target character}
}
\value{
replaced character
}
\description{
Replace zenkaku-symbol by hankaku-symbol
}
