% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{str_to_han}
\alias{str_to_han}
\title{Replace zenkaku alphabet, number and symbols by hankaku}
\usage{
str_to_han(x)
}
\arguments{
\item{x}{target character}
}
\value{
replaced character
}
\description{
Replace zenkaku alphabet, number and symbols by hankaku
}
