% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{str_to_ymd}
\alias{str_to_ymd}
\title{transform string to ymd data table.}
\usage{
str_to_ymd(
  str,
  Date.beg = NULL,
  Date.end = NULL,
  use.jpera = TRUE,
  use.exceldate = FALSE,
  exceldate.origin = as.Date("1904-01-01")
)
}
\arguments{
\item{str}{target character or sequence of character.}

\item{Date.beg}{potentially start date; used for validation and estimation.}

\item{Date.end}{potentially end date; used for validation and estimation.}

\item{use.jpera}{Allow Japanese era case (e.g., R2.4.5 or 令和3年4月11日).}

\item{use.exceldate}{Allow to check the value which is originally excel date but occationally transformed to normal value.}

\item{exceldate.origin}{Origin of excel date. In default, it is 1900-01-01 if the file is originally created on Windows and 1904-01-01 on Mac.}
}
\value{
transformed data.frame
}
\description{
transform string to ymd data table like data.frame(y=2022,m=5,d=12). NA will be filled for uncertain elements.
}
