% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{strong_require}
\alias{strong_require}
\title{load package with trying package install when it is not found.}
\usage{
strong_require(pck, location = NULL)
}
\arguments{
\item{pck}{character of package name}

\item{location}{if location is "github.com/xxx", the package is installed from github.com}
}
\value{
True: found and loaded, False: fail to load
}
\description{
Try load package by require. If the package does not exist, install.package or install_github is tried.
}
