% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{throw}
\alias{throw}
\title{Throw exception}
\usage{
throw(message, IgnoreThrower = FALSE)
}
\arguments{
\item{message}{message of exception}

\item{IgnoreThrower}{Logical: if true, the callstack data don't include the function it call throw.}
}
\description{
Throw exception with message.
}
