% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_diagram.R
\name{to_triangle}
\alias{to_triangle}
\title{Change axis for triangle diagram.}
\usage{
to_triangle(x, y)
}
\arguments{
\item{x}{position of first element.}

\item{y}{position of second element. Note that the third element is not required because z is always 1-x-y.}
}
\description{
Change axis for triangle diagram.
}
