% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_diagram.R
\name{triangle_lines}
\alias{triangle_lines}
\title{Plot lines on triangle diagram.}
\usage{
triangle_lines(x, y, ...)
}
\arguments{
\item{x}{position of first element.}

\item{y}{position of second element. Note that the third element is not required because z is always 1-x-y.}

\item{...}{Other parameters for lines function.}
}
\description{
Plot lines on triangle diagram.
}
