% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_diagram.R
\name{triangle_plot}
\alias{triangle_plot}
\title{Plot for triangle diagram.}
\usage{
triangle_plot(xlab = "", ylab = "", zlab = "")
}
\arguments{
\item{xlab}{label of x axis.}

\item{ylab}{label of y axis.}

\item{zlab}{label of z axis.}
}
\description{
Plot for triangle diagram.
}
