% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_diagram.R
\name{triangle_points}
\alias{triangle_points}
\title{Plot points on triangle diagram.}
\usage{
triangle_points(x, y, ...)
}
\arguments{
\item{x}{position of first element.}

\item{y}{position of second element. Note that the third element is not required because z is always 1-x-y.}

\item{...}{Other parameters for points function.}
}
\description{
Plot points on triangle diagram.
}
