% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_diagram.R
\name{triangle_text}
\alias{triangle_text}
\title{Plot text on triangle diagram.}
\usage{
triangle_text(x, y, ...)
}
\arguments{
\item{x}{position of first element.}

\item{y}{position of second element. Note that the third element is not required because z is always 1-x-y.}

\item{...}{Other parameters for text function.}
}
\description{
Plot text on triangle diagram.
}
