% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_diagram.R
\name{triangle_vectormap}
\alias{triangle_vectormap}
\title{Plot triangle vector map.}
\usage{
triangle_vectormap(x, y, vx, vy, xlab = "", ylab = "", zlab = "",
  vlim = NA, arrow.len = 0.05, arrow.headlen = 0.1,
  col = rev(gray.colors(20)))
}
\arguments{
\item{x}{position of first element.}

\item{y}{position of second element. Note that the third element is not required because z is always 1-x-y.}

\item{vx}{vector/allow of first element.}

\item{vy}{vector/allow of second element. Note that the third element is not required because z is always 1-x-y.}
}
\description{
Plot triangle vector map.
}
