% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{try_}
\alias{try_}
\title{Try execute function which might throw exception.}
\usage{
try_(expr, silent = TRUE)
}
\arguments{
\item{expr}{Function which should be executed.}

\item{silent}{Logical value: if TRUE, the error message is not shown.}
}
\value{
return value if expr don't throw, or exception if it throw.
}
\description{
Try execute function which might throw exception.
}
