% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{try_catch}
\alias{try_catch}
\title{Try excute function which might throw exception.}
\usage{
try_catch(expr, silent = TRUE)
}
\arguments{
\item{expr}{Function which should be excuted.}

\item{silent}{Logical value: if TRUE, the error message is not shown.}
}
\value{
return value if expr don't throw, or exception if it throw.
}
\description{
Try excute function which might throw exception.
}
