% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{try_library}
\alias{try_library}
\title{Read package after install if it is required}
\usage{
try_library(Names)
}
\arguments{
\item{Name}{Name of the package}
}
\description{
Read package after install if it is required
}
