% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{unescape_md}
\alias{unescape_md}
\title{Unescape all markdown character.}
\usage{
unescape_md(str)
}
\arguments{
\item{str}{character vector}
}
\value{
Returns a character vector.
}
\description{
Unescape all markdown character.
}
