% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{unescape_regex}
\alias{unescape_regex}
\title{Unescape all regex character.}
\usage{
unescape_regex(str)
}
\arguments{
\item{str}{character vector}
}
\value{
Returns a character vector.
}
\description{
Unescape all regex character.
}
