% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_field.R
\name{vector_field.color}
\alias{vector_field.color}
\title{Plot vector field for 2D. Color is the strength of the vector.}
\usage{
vector_field.color(
  x_seq,
  y_seq,
  dx,
  dy,
  colrate,
  arrow.length = NA,
  arrow.sublength = 0.25,
  xyrate = 1,
  col = rev(heat.colors(100)),
  limit = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{x_seq}{x-axix}

\item{y_seq}{y-axix}

\item{dx}{matrix of selection gradient for x}

\item{dy}{matrix of selection gradient for y}

\item{colrate}{influence of the strength on the color.}

\item{arrow.length}{Basic length of the arrow.}

\item{arrow.sublength}{Basic sub length of the arrow.}

\item{xyrate}{relative length of x and y values.}

\item{col}{color of the vector}

\item{limit}{limit function with two argument x,y. If limit(x,y) return true, the arrow is not drown.}

\item{add}{boolian: whether add existign plot space or not.}

\item{...}{same with the plot function.}
}
\description{
Plot vector field for 2D. Color is the strength of the vector.
}
