% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_field.R
\name{vector_field}
\alias{vector_field}
\title{Plot vector field for 2D. Length is the strength of the vector.}
\usage{
vector_field(
  x_seq,
  y_seq,
  dx,
  dy,
  arrow.length,
  arrow.sublength,
  col = "black",
  limit = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{x_seq}{x-axix}

\item{y_seq}{y-axix}

\item{dx}{matrix of selection gradient for x}

\item{dy}{matrix of selection gradient for y}

\item{arrow.length}{Basic length of the arrow.}

\item{arrow.sublength}{Basic sub length of the arrow.}

\item{col}{color of the vector}

\item{limit}{limit function with two argument x,y. If limit(x,y) return true, the arrow is not drown.}

\item{add}{boolian: whether add existign plot space or not.}

\item{...}{same with the plot function.}
}
\description{
Plot vector field for 2D. Length is the strength of the vector.
}
