% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{warnings_message}
\alias{warnings_message}
\title{Create warning message from warnings data}
\usage{
warnings_message(message)
}
\arguments{
\item{message}{Header mesasge}
}
\value{
message with warnings data
}
\description{
Create warning message which include the information of warnings data.
}
