% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{weekno_to_date}
\alias{weekno_to_date}
\title{Calculate date range from week number}
\usage{
weekno_to_date(wno)
}
\arguments{
\item{wno}{the week number of the focal week}
}
\value{
data frame with from (first date) and to (last date)
}
\description{
Calculate date range of the focal week number.
}
