% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{weekno_to_firstdate}
\alias{weekno_to_firstdate}
\title{Calculate first date from week number}
\usage{
weekno_to_firstdate(wno)
}
\arguments{
\item{wno}{the week number of the focal week}
}
\value{
first date
}
\description{
Calculate first date of the focal week number.
}
