% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{ymd_eq_greater}
\alias{ymd_eq_greater}
\title{compare two ymd by equal_or_greater}
\usage{
ymd_eq_greater(ymd1, ymd2)
}
\arguments{
\item{ymd1}{l.f.s.}

\item{ymd2}{r.h.s.}
}
\value{
ymd1>=ymd2
}
\description{
compare two ymd by equal_or_greater
}
