% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{ymd_greater}
\alias{ymd_greater}
\title{compare two ymd by greater}
\usage{
ymd_greater(ymd1, ymd2)
}
\arguments{
\item{ymd1}{l.f.s.}

\item{ymd2}{r.h.s.}
}
\value{
ymd1>ymd2
}
\description{
compare two ymd by greater
}
