% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{ymd_include}
\alias{ymd_include}
\title{check whether given ymd is included in the given range}
\usage{
ymd_include(ymd, ymd_from, ymd_to)
}
\arguments{
\item{ymd}{target ymd}

\item{ymd_from}{start of the range}

\item{ymd_to}{end of the range}
}
\value{
ymd_begin <= ymd & ymd < ymd_end
}
\description{
compare two ymd by equal
}
