% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{ymd_less}
\alias{ymd_less}
\title{compare two ymd by less}
\usage{
ymd_less(ymd1, ymd2)
}
\arguments{
\item{ymd1}{l.f.s.}

\item{ymd2}{r.h.s.}
}
\value{
ymd1<ymd2
}
\description{
compare two ymd by less
}
