% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{ymd_to_datestr}
\alias{ymd_to_datestr}
\title{transform ymd data table to string.}
\usage{
ymd_to_datestr(
  ymd,
  dateformat = "\%Y.\%M.\%D",
  range_sep = "-",
  na.fill = "?",
  mode = "full"
)
}
\arguments{
\item{ymd}{data.frame with column y, m and d for year, month and day.}

\item{dateformat}{output format with Y, y, M, m, D, d. See detais.}

\item{range_sep}{separation of range if ymd is range format.}

\item{na.fill}{filled character for missing element.}

\item{mode}{determining data filling mode = c("full","maximum","compact","short","minimum"). see detail.}
}
\value{
transformed str
}
\description{
transform ymd data table like data.frame(y=2022,m=5,d=12) to string. NA will be filled by na.fill.
}
\details{
format of each mode is like the following.
	maximum: always range mode e.g., 2012.03.04-2012.03.04 2023.04.05-2023.05.06
	full (default): range mode is used only when beg and end is different e.g., 2012.03.04 2023.04.05-2023.05.06
	compact: range mode remove duplicating information e.g., 2023.04.05-05.06 2023.04.??
	minimum: (information loss can occur depending on dateformat) remove NA and duplicating info e.g., 2023.04-05
dateformat can use the following special characters.
	(percent)Yj(at): NOT WORKING
	(percent)YJ(at): NOT WORKING
	(percent)Y(at): YYYY年
	(percent)y(at): YY年
	(percent)M(at): MM月
	(percent)m(at): M月
	(percent)D(at): DD日
	(percent)d(at): D日
	(percent)Y: YYYY
	(percent)y: YY
	(percent)M: MM
	(percent)m: M
	(percent)D: DD
	(percent)d: D
}
