% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{ymd_to_str}
\alias{ymd_to_str}
\title{transform ymd data table to string.}
\usage{
ymd_to_str(ymd, format = "\%Y.\%m.\%d", na.fill = "?")
}
\arguments{
\item{ymd}{data.frame with column y, m and d for year, month and day.}

\item{format}{output format with y,Y,m and d.}

\item{na.fill}{filled character for missing element.}
}
\value{
transformed str
}
\description{
transform ymd data table like data.frame(y=2022,m=5,d=12) to string. NA will be filled by na.fill.
}
