% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{ymd_to_weekno}
\alias{ymd_to_weekno}
\title{Calculate weekno from Year, month and Day}
\usage{
ymd_to_weekno(year, month, day)
}
\arguments{
\item{year}{the year of the focal date}

\item{month}{the month of the focal date}

\item{day}{the date of the focal date}
}
\value{
Week number
}
\description{
Calculate weekno from Year, month and Day
}
