% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week.R
\name{yw_to_weekno}
\alias{yw_to_weekno}
\title{Calculate weekno from year and week number on the year.}
\usage{
yw_to_weekno(year, week)
}
\arguments{
\item{year}{the year of the focal week}

\item{week}{the week of the focal week}
}
\value{
Week number
}
\description{
Calculate weekno from Date
}
