% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.at}
\alias{file.at}
\title{Return function which generate standardized path of the focal directory}
\usage{
file.at(filedir)
}
\arguments{
\item{filedir}{file directory}
}
\value{
function of standardized path generator
}
\description{
Returned function generates standardized path of the given directory, base_dir.
}
