% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file.fout}
\alias{file.fout}
\title{generate file output}
\usage{
file.fout(filepath, encoding = getOption("encoding"), raw = FALSE)
}
\arguments{
\item{filepath}{file path (character)}

\item{encoding}{encoding}

\item{raw}{raw option}
}
\value{
fout function
}
\description{
Generate file output
}
