% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{match.order}
\alias{match.order}
\title{<Deprecated> Return index numbers of the target sequence which is matching to the obj,}
\usage{
match.order(from, trg)
}
\arguments{
\item{from}{Value, vector or matrix}

\item{trg}{Target sequence}
}
\value{
Found once: Index numbers, Found multiple times: NA, Didn't find: 0
}
\description{
Find index number at which trg is equal to obj. Return 0 if it is not found in trg, and return NA if it has been multitmes found.
}
