% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{source_rscript}
\alias{source_rscript}
\title{read and run R script code from a file, with changing directory.}
\usage{
source_rscript(
  file,
  args = NULL,
  local = parent.env(.GlobalEnv),
  ...,
  chdir = TRUE
)
}
\arguments{
\item{file}{file name}

\item{args}{set Rscript arguments}

\item{local}{local environment; default is baseenv}

\item{...}{argument for source}

\item{chdir}{change dir}
}
\description{
read and run R script code from a file, with changing directory.
}
