% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{start_ownself}
\alias{start_ownself}
\title{Execute ownself in parallele with arguments.}
\usage{
start_ownself(ownpath, ...)
}
\arguments{
\item{ownpath}{Own file path}

\item{...}{optional arguments for rscript file.}
}
\description{
Run ownself as Rscript in parallele with arguments. In parallele started files, this function does not start new process.
}
