% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_field.R
\name{vector_field}
\alias{vector_field}
\title{Plot vector field for 2D. Length is the strength of the vector.}
\usage{
vector_field(dxf, dyf, arrow.length, arrow.sublength, xlim = c(0, 1),
  ylim = c(0, 1), num = 20, col = "black", limit = NULL, add = FALSE,
  ...)
}
\arguments{
\item{dxf}{function for the gradient for x-axix}

\item{dyf}{function for the gradient for y-axix}

\item{arrow.length}{Basic length of the arrow.}

\item{arrow.sublength}{Basic sub length of the arrow.}

\item{num}{Density of the arrows.}

\item{...}{same with the plot function.}
}
\description{
Plot vector field for 2D. Length is the strength of the vector.
}
